app/requirements.txt
Flask==2.3.2
SQLAlchemy==1.4.52
Pillow==10.0.1
requests==2.31.0
loguru==0.7.0
exifread==3.0.0
Werkzeug==2.3.6
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
